Hide Volumepanel on incoming call
When receiving an incoming phone call and the user recently
changed the volume on the phone, the volume dialog overlaps
the heads up dialog for incoming call. This prevents the
user to interacting with it. This change dismiss the volume dialog
if it is showing when receiving an incoming call.

Change-Id: I3a9fd9376cd8b4334e65765bd2eda9a7146d808b